move NoPricesBanner up one folder
